langgraph_project/tools/midjourney.py
import os
from langchain_core.tools import tool
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# This is a placeholder for a real Midjourney API client.
# In a real scenario, you would use a library or directly call the Midjourney API.
# You might need an API key, e.g., os.getenv("MIDJOURNEY_API_KEY")

@tool
def midjourney_image_generator(prompt: str) -> str:
    """
    Generates an image using a conceptual Midjourney API based on the provided prompt.
    This is a mock tool and returns a placeholder image URL.
    """
    if not prompt:
        return "Error: Prompt cannot be empty for Midjourney image generation."

    print(f"🎨 [Midjourney Tool] Received request to generate image for prompt: '{prompt}'")
    
    # Simulate API call and image generation
    # In a real implementation:
    # api_key = os.getenv("MIDJOURNEY_API_KEY")
    # client = MidjourneyClient(api_key=api_key)
    # image_url = client.generate(prompt)
    
    mock_image_url = f"https://example.com/mock_image_{prompt.lower().replace(' ', '_')}.png"
    response_message = f"Successfully generated mock image for '{prompt}'. URL: {mock_image_url}"
    
    print(f"🖼️ [Midjourney Tool] {response_message}")
    return response_message
