pyproject.toml
[tool.poetry]
name = "langgraph-midjourney-project"
version = "0.1.0"
description = "A LangGraph project with a Midjourney tool"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.0"
langgraph = "^0.0.30"
langchain_openai = "^0.1.0"
python-dotenv = "^1.0.0"
tavily-python = "^0.3.0" # Often useful for web search capabilities

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
